Decouple the grid from the tiler
This enables grids other than `GlobalGeodetic` to be associated with a
`GDALTiler`, allowing tilesets in different grid systems to be
generated.

To capitalise on this the `terrain-build` and `terrain-extents` tools
have the new `--profile` and `--tile-size` options.  `terrain-build`
also has the `--output-format` option allowing GDAL formats to be
generated in addition to terrain tiles. This turns the tool into a
general purpose tile generator.