Add a blank terrain tile for use as a root tile in a tileset
Unless the GDAL dataset passed to `ctb-tile` crosses the Prime Meridian one of
either of the root tiles `/0/0/0.terrain` or '/0/0/1.terrain` will be missing.
However, both tiles are required by CesiumJS.  The added tile specifies no child
tiles and can be manually added to the appropriate root tile in a tileset in
order to resolve the problem. See issue #1 for details.