Fix compilation with external gdal library
Checking for libgdal2 was failing when the library was not in the
system library path. Additionally, the pthread lib was missing from the
link command.

With this commit, it is now possible to install libgdal2 in a user
directory and compile ctb with this version.

For example:
cmake -DGDAL_LIBRARY_DIR=/home/gberaudo/dev/inst/lib
-DGDAL_LIBRARY=/home/gberaudo/dev/inst/lib/libgdal.so
-DGDAL_INCLUDE_DIR=/home/gberaudo/dev/inst/include ..